requirements.txt
# Core dependencies
bleak>=0.21.0              # Modern async BLE library 
paho-mqtt>=1.6.0           # MQTT client
pydantic>=2.0.0            # Configuration validation
pyyaml>=6.0                # YAML configuration parsing
asyncio-mqtt>=0.13.0       # Async MQTT client wrapper

# Optional monitoring and web features  
aiohttp>=3.8.0             # Web dashboard and health checks
prometheus-client>=0.17.0   # Metrics export

# Development and testing
pytest>=7.0.0              # Testing framework
pytest-asyncio>=0.21.0     # Async test support
black>=23.0.0              # Code formatting
pylint>=2.17.0             # Linting
mypy>=1.0.0                # Type checking